fix(toolbox) don't mix web and native actions on the same file
