+++ react/features/toolbox/actions.web.js
@@ -5,20 +5,20 @@ import type { Dispatch } from 'redux';
 import { isLayoutTileView } from '../video-layout';
 
 import {
+    CLEAR_TOOLBOX_TIMEOUT,
     FULL_SCREEN_CHANGED,
     SET_FULL_SCREEN,
-    SET_OVERFLOW_DRAWER
+    SET_OVERFLOW_DRAWER,
+    SET_OVERFLOW_MENU_VISIBLE,
+    SET_TOOLBAR_HOVERED,
+    SET_TOOLBOX_TIMEOUT,
+    SET_TOOLBOX_TIMEOUT_MS
 } from './actionTypes';
-import {
-    clearToolboxTimeout,
-    setToolboxTimeout,
-    setToolboxTimeoutMS,
-    setToolboxVisible
-} from './actions.native';
+import { setToolboxVisible } from './actions.any';
 
 declare var interfaceConfig: Object;
 
-export * from './actions.native';
+export * from './actions.any';
 
 /**
  * Docks/undocks the Toolbox.
@@ -180,9 +180,88 @@ export function hideToolboxOnTileView() {
         const state = getState();
         const { overflowDrawer } = state['features/toolbox'];
 
-
         if (!overflowDrawer && isLayoutTileView(state)) {
             dispatch(hideToolbox(true));
         }
+    };
+}
+
+/**
+ * Signals that toolbox timeout should be cleared.
+ *
+ * @returns {{
+ *     type: CLEAR_TOOLBOX_TIMEOUT
+ * }}
+ */
+export function clearToolboxTimeout(): Object {
+    return {
+        type: CLEAR_TOOLBOX_TIMEOUT
+    };
+}
+
+/**
+ * Shows/hides the overflow menu.
+ *
+ * @param {boolean} visible - True to show it or false to hide it.
+ * @returns {{
+ *     type: SET_OVERFLOW_MENU_VISIBLE,
+ *     visible: boolean
+ * }}
+ */
+export function setOverflowMenuVisible(visible: boolean): Object {
+    return {
+        type: SET_OVERFLOW_MENU_VISIBLE,
+        visible
+    };
+}
+
+/**
+ * Signals that toolbar is hovered value should be changed.
+ *
+ * @param {boolean} hovered - Flag showing whether toolbar is hovered.
+ * @returns {{
+ *     type: SET_TOOLBAR_HOVERED,
+ *     hovered: boolean
+ * }}
+ */
+export function setToolbarHovered(hovered: boolean): Object {
+    return {
+        type: SET_TOOLBAR_HOVERED,
+        hovered
+    };
+}
+
+/**
+ * Dispatches an action which sets new timeout and clears the previous one.
+ *
+ * @param {Function} handler - Function to be invoked after the timeout.
+ * @param {number} timeoutMS - Delay.
+ * @returns {{
+ *     type: SET_TOOLBOX_TIMEOUT,
+ *     handler: Function,
+ *     timeoutMS: number
+ * }}
+ */
+export function setToolboxTimeout(handler: Function, timeoutMS: number): Object {
+    return {
+        type: SET_TOOLBOX_TIMEOUT,
+        handler,
+        timeoutMS
+    };
+}
+
+/**
+ * Dispatches an action which sets new toolbox timeout value.
+ *
+ * @param {number} timeoutMS - Delay.
+ * @returns {{
+ *     type: SET_TOOLBOX_TIMEOUT_MS,
+ *     timeoutMS: number
+ * }}
+ */
+export function setToolboxTimeoutMS(timeoutMS: number): Object {
+    return {
+        type: SET_TOOLBOX_TIMEOUT_MS,
+        timeoutMS
     };
 }
